requirements.txt
websocket-client
urwid
